==> R/02.general_definitions.R <==
## this file contains some definitions that are prevalent throughtout the project
## mainly regarding colors and orders pf dates and phenological stages.

## whenever possible we will be using "Set1"
colors_of_set1 <- c("#E41A1C", "#377EB8", "#4DAF4A", "#984EA3", "#FF7F00", "gold", "#A65628", "#F781BF", "#999999")

positions_ps <- c("White buds", "Initial bloom", "Full bloom", "Initial petal fall", "Complete petal fall", "Leaf")
positions_cps <- c("Spadona White buds", "Spadona Initial bloom", "Coscia White buds", "Coscia Initial bloom", "Coscia Full bloom", "Coscia Initial petal fall", "Coscia Complete petal fall", "Spadona Leaf", "Coscia Leaf")


color_values_cps_init <- c(
  "Spadona White buds" = "#E41A1C",
  "Spadona Initial bloom" = "#FF7F00",
  "Coscia White buds" = "#377EB8",
  "Coscia Initial bloom" = "#4DAF4A"
)
color_values_cps <- c(
  "Spadona White buds" = "#E41A1C",
  "Spadona Initial bloom" = "#FF7F00",
  "Coscia White buds" = "#377EB8",
  "Coscia Initial bloom" = "#4DAF4A",
  "Coscia Full bloom" = "#A65628",
  "Coscia Initial petal fall" = "#984EA3",
  "Coscia Complete petal fall" = "#F781BF",
  "Spadona Leaves" = "#999999",
  "Coscia Leaves" = "gold"
)
color_values_cd <- c(
  "24.03.2019" = "#E41A1C",
  "27.03.2019" = "#FF7F00",
  "31.03.2019" = "#FFD700",
  "11.04.2019" = "#377EB8",
  "13.04.2019" = "#4DAF4A",
  "15.04.2019" = "#A65628",
  "18.04.2019" = "#984EA3",
  "21.04.2019" = "#F781BF"
)

color_values_ps <- c(
  "White buds" = "#E41A1C",
  "Initial bloom" = "#377EB8",
  "Full bloom" = "gold",
  "Initial petal fall" = "#984EA3",
  "Complete petal fall" = "#FF7F00",
  "Leaf" = "#4DAF4A"
)


positions_cd <- c("24.03.2019", "27.03.2019", "31.03.2019", "11.04.2019", "13.04.2019", "15.04.2019", "18.04.2019", "21.04.2019")
positions_cd_initial <- c("24.03.2019", "27.03.2019", "31.03.2019", "11.04.2019", "13.04.2019")
positions_cd_s <- c("24.03", "27.03", "31.03", "11.04", "13.04", "15.04", "18.04", "21.04")
color_values_cd_s <- c(
  "24.03" = "#E41A1C",
  "27.03" = "#FF7F00",
  "31.03" = "gold",
  "11.04" = "#377EB8",
  "13.04" = "#4DAF4A",
  "15.04" = "#A65628",
  "18.04" = "#984EA3",
  "21.04" = "#F781BF"
)
